Trim/move ad-free things for better organisation and performance
